Wrote a function to create all the Divs
Rather than writing them all out i realized there was too much code for only a few circles so I remembered a lesson on looping to create random art, and then created a loop to create divs and saved a ton of space and time in the code. I then  altered the css while testing the box sizes that will fit the drawing area.